move the album/song getters for album/song relationship to new AlbumLibrary helper class
