+++ src/main/java/app/util/AlbumLibrary.java
@@ -1,0 +1,40 @@
+package app.util;
+
+import app.player.Album;
+import app.player.Song;
+
+import java.util.List;
+
+public class AlbumLibrary {
+
+    private final List<Album> albums;
+
+    public AlbumLibrary(List<Album> albums) {
+        this.albums = albums;
+    }
+
+    public Album getAlbumByName(String name) {
+        for (Album album : albums) {
+            if (album.getName().equals(name)) return album;
+        }
+        return null;
+    }
+
+    public Album getAlbumForSong(String songTitle) {
+        for (Album album : albums) {
+            for (Song song : album.getSongs()) {
+                if (song.getTitle().equals(songTitle)) return album;
+            }
+        }
+        return null;
+    }
+
+    public Song getSongByTitle(String songTitle) {
+        for (Album album : albums) {
+            for (Song song : album.getSongs()) {
+                if (song.getTitle().equals(songTitle)) return song;
+            }
+        }
+        return null;
+    }
+}
